implement dll_thunk.h .
To get rid of many clang -Weverything warnings about an unused macro.

+++ fc-solve/source/internal_move_struct.h
@@ -90,7 +90,7 @@ typedef fcs_move_t fcs_internal_move_t;
 typedef struct
 {
     fcs_internal_move_t * moves;
-    unsigned short num_moves;
+    size_t num_moves;
 } fcs_move_stack_t;
 
 #ifdef __cplusplus
